sam3-server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
segment-geospatial>=0.11.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.0
groundingdino-py>=0.4.0
pydantic>=2.0.0
